1.2 Update: /gamemode, /kill, bug fixes & chat messages
Added /gamemode <survival/creative/spectator/adventure> <player>, as well as /gmc, /gms, /gmsp, & /gma aliases. If a <player> is not set it defaults to yourself.

Added /kill <player>. If a <player> is not specified it defaults to yourself.

Fixed bugs with internal errors on running certain commands in console, or setting <player> to a player that doesnt exist

Remade Join & Leave messages